requirements.txt
aiogram==3.20.0.post0
python-dotenv==1.1.0
torch==2.5.1+cu121
transformers==4.51.3
